Add files via upload
In the past few decades, there has been an unprecedented interest in resilience studies across natural science and engineering disciplines. This has resulted in numerous frameworks that are highly context context-specific, with each framework based on a set of rules and assumptions, thus compromising the generality of its application. Our study focuses on the inherent traits of a system that aid in managing disturbances, thus deviating from the traditional mathematical frameworks in resilience studies. We identify two important traits of a system, which either provide capabilities of withstanding disturbances or the ability to recover from these disturbances.  A spring-damper mechanical system (SDMS) provides a suitable analogy, as the spring and damper represent recovery and withstanding capabilities respectively to an external force. We test the applicability of SDMS to represent ecological/biological systems by inducing bi-stability into the SDMS equation and develop an integrated resilience index (IRI) that can distinguish systems based on their resilient properties. We examine the applicability of IRI in two scenarios : 1) systems with differential resilience properties experiencing similar external disturbances and 2) systems with differential resilience properties experiencing varying intensity of external disturbances. Numerical simulations were conducted to examine SDMSs and mathematical models representing ecological systems for these scenarios. IRI was successful in distinguishing systems based on their resilient properties. Importantly, the formulation of IRI is based on the long-term response of the system to a continuously varying external disturbance and thus, overcomes limitations of the traditional resilience frameworks, particularly in the context of engineering and ecological resilience frameworks applied in ecology. Although the modelling examples are motivated by ecological or biological systems, this novel approach to measure resilience is simple and versatile, thus opening opportunities to be tested on empirical datasets of diverse disciplines.